+++ Engine/Base/Buffer/VertIdxBuff.h
@@ -27,7 +27,7 @@ private:
 	void SetResDesc(UINT size);
 
 protected:
-	void Initialize(uint32_t sizeVB, const std::vector<uint32_t>& indices);
+	void Initialize(uint32_t sizeVB, const std::vector<uint16_t>& indices);
 	void Initialize(uint32_t sizeVB);
 	void BuffInitialize(ID3D12Device* dev, UINT sizeVB, UINT sizeIB = NULL, uint16_t* indices = nullptr, int indicesSize = NULL);
 	void IASetVertIdxBuff();
@@ -37,7 +37,7 @@ public:
 	virtual ~VertIdxBuff() = default;
 
 	virtual void SetVertices() = 0;
-	void SetIndices(const std::vector<uint32_t>& indices);
+	void SetIndices(const std::vector<uint16_t>& indices);
 
 	//	Getter
 	const D3D12_RESOURCE_DESC& GetResDesc() { return resDesc_; }
